Fixed Simulation
fixed the simulation now works and diplays the  number of wins in the console was not sure how you were hooking up the combat log i seen the script and where you were trying to set it in the code. but was not attached to anything not sure where you wanted it we can talk about it in class thursday.  but this should be enogh to show him something on thurday